Move common code to AntWithFindResourceClassLoader.
