+++ core/src/main/java/hudson/PluginFirstClassLoader.java
@@ -27,16 +27,13 @@ import java.io.Closeable;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.lang.reflect.Field;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Enumeration;
 import java.util.List;
-import java.util.Vector;
 
-import org.apache.tools.ant.AntClassLoader;
-import org.apache.tools.ant.Project;
+import jenkins.util.AntWithFindResourceClassLoader;
 
 /**
  * classLoader which use first /WEB-INF/lib/*.jar and /WEB-INF/classes before core classLoader
@@ -45,22 +42,12 @@ import org.apache.tools.ant.Project;
  * @since 1.371
  */
 public class PluginFirstClassLoader
-    extends AntClassLoader
+    extends AntWithFindResourceClassLoader
     implements Closeable
 {
 
-    private final Vector pathComponents;
-
     public PluginFirstClassLoader() {
-        super();
-
-        try {
-            Field $pathComponents = AntClassLoader.class.getDeclaredField("pathComponents");
-            $pathComponents.setAccessible(true);
-            pathComponents = (Vector)$pathComponents.get(this);
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            throw new Error(e);
-        }
+        super(null, false);
     }
 
     private List<URL> urls = new ArrayList<URL>();
@@ -117,27 +104,5 @@ public class PluginFirstClassLoader
     {
         InputStream is = super.getResourceAsStream( name );
         return is;
-    }
-
-    /**
-     * As of 1.8.0, {@link AntClassLoader} doesn't implement {@link #findResource(String)}
-     * in any meaningful way, which breaks fast lookup. Implement it properly.
-     */
-    @Override
-    protected URL findResource(String name) {
-        URL url = null;
-
-        // try and load from this loader if the parent either didn't find
-        // it or wasn't consulted.
-        Enumeration e = pathComponents.elements();
-        while (e.hasMoreElements() && url == null) {
-            File pathComponent = (File) e.nextElement();
-            url = getResourceURL(pathComponent, name);
-            if (url != null) {
-                log("Resource " + name + " loaded from ant loader", Project.MSG_DEBUG);
-            }
-        }
-
-        return url;
     }
 }
